chore: move hardcoded template data to scripts/ (seed-only)
