+++ scripts/map-template-data.ts
